tracker-poetry/pyproject.toml
[project]
name = "tracker-poetry"
version = "0.1.0"
description = ""
authors = [
    {name = "IvanRamosDataTech",email = "ivanramos.datatech@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pendulum (>=3.0.0,<4.0.0)",
    "sqlquery (>=1.1.0,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "xlsxwriter (>=3.2.2,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "msal (>=1.32.3,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
]

[tool.poetry]
packages = [{include = "tracker_poetry", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
